pyproject.toml
[tool.poetry]
name = "health-manager"
version = "0.1.0"
description = ""
authors = ["Arnab Mukherjee <am0072008@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
hm_app = "health_manager.server:run"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
httpx = "^0.28.1"
pytest = "^8.3.5"
sqlmodel = "^0.0.24"
sqlalchemy = "^2.0.40"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
